ocr/requirements.txt

azure-cognitiveservices-vision-computervision
pillow